build(ci) Github action configuration and dockerfile
Create a simple CI pipeline with Github actions. The pipeline is to
build with make and static analysis on each commit. Please note since
the project Makefile is still empty the building will be added in the
sudsquent commits. Each pull request must pass the two steps before it's
merged. Also includes the dockerfile used to create the docker
contatiner, the action runs inside.