feat(drivers) IR remote driver to decode IR signals
Sumo robot must be able to decode the start signal sent at every start
of the match. the signal is sent by the IR transmitter and picked up by
the IR receiver on the board. Implement the driver that decodes the
signal by measuring the time between the pulses on GPIO pin connected to
the IR receiver. Besides the start signal use the functionality to
control the robot remotely.